Fixed warnings and 32 bit compilation
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/cinematiccamera/branches/skybox@67 48bde7f2-199b-42c5-ada7-780c78a47d76

+++ src/CinematicView.cpp
@@ -67,7 +67,7 @@ void megamol::cinematiccamera::CinematicView::Render(const mmcRenderViewContext&
 	}
 	else {
 
-		kfc->setTimeofKeyframeToGet(context.Time);
+		kfc->setTimeofKeyframeToGet(static_cast<float>(context.Time));
 		if ((*kfc)(CallCinematicCamera::CallForGetKeyframeAtTime)){
 			kfc->getInterpolatedKeyframe().putCamParameters(this->cam.Parameters());
